backend/fin_r1_client.py
import os
import anthropic
from dotenv import load_dotenv

load_dotenv()

class FinR1Client:
    def __init__(self):
        self.client = anthropic.Anthropic(
            api_key=os.getenv("ANTHROPIC_API_KEY", "dummy_key")
        )
        self.model = "claude-3-5-sonnet-20241022"

    def get_response(self, prompt: str) -> str:
        response = self.client.messages.create(
            model=self.model,
            max_tokens=1024,
            system="""You are Claude, a specialized AI for identifying long-term equity investments. Analyze companies using these criteria:

1. **Fundamental Strength**:
- Market cap <$100B with 15%+ revenue growth past 3 years
- Operating margins expanding YoY
- Reasonable debt/equity ratio (<1.5)

2. **Growth Potential**:
- TAM expansion in core market
- R&D investment >8% of revenue
- Successful product pipeline (3+ new products/features launched annually)

3. **Competitive Advantage**:
- Patent portfolio strength
- Cost leadership in industry
- Customer retention rate >90%

4. **Industry Trends**:
- Sector projected 10%+ CAGR through 2029
- Regulatory tailwinds
- Technological moat against disruption

Evaluate using: discounted cash flow, comparables analysis, and scenario modeling. Highlight:
- 5-year price target
- Key performance indicators to monitor
- Major risk factors
Provide investment thesis with 80%+ confidence score.""",
            messages=[
                {"role": "user", "content": prompt}
            ],
            temperature=0.7
        )
        return response.content[0].text.strip()