backend/requirements.txt
blake3==1.0.5
jiter==0.10.0
networkx==3.5
opencv-python-headless>=4.11.0
contourpy==1.3.2
tiktoken==0.9.0
deprecation==2.1.0
hf-xet==1.1.5
httpx>=0.15.0,<=0.27.0
matplotlib==3.10.3
Jinja2==3.1.6
psutil==7.0.0
peewee==3.18.2
websocket-client==1.8.0
mlx==0.26.3
attrs>=22.2.0
python-json-logger==3.3.0
frozenlist==1.7.0
aiosignal==1.4.0
cloudpickle==3.1.1
pluggy==1.6.0
typer==0.16.0
einops==0.8.1
jsonschema==4.24.0
transformers==4.53.1
huggingface-hub==0.33.2
openai==1.90.0
beautifulsoup4==4.13.4
rich-toolkit==0.14.8
python-dotenv==1.1.1
llvmlite==0.44.0
multiprocess==0.70.16
filelock==3.18.0
prometheus-fastapi-instrumentator==7.1.0
seaborn==0.13.2
msgpack==1.0.3
protobuf<5.0.0,>=3.19.5
alpaca-py>=0.28.0
vaderSentiment==3.3.2
shellingham==1.5.4
alpaca-trade-api>=3.0.0
PyYAML==6.0.1
yfinance==0.2.58
torchaudio==2.7.0
py-cpuinfo==9.0.0
pytest==8.4.1
rich>=13.7.1
interegular==0.3.3
ninja==1.11.1.4
safetensors==0.5.3
dill==0.3.8
markdown-it-py==3.0.0
idna==3.10
xgrammar==0.1.19
llguidance==0.7.30
aiohttp==3.12.13
pycparser==2.22
cachetools<6.0,>=5.5.0
ruff==0.12.2
multitasking==0.0.11
requests>=2.32.2
depyf==0.18.0
msgspec==0.19.0
urllib3==1.26.20
dnspython==2.7.0
setuptools==79.0.1
charset-normalizer==2.1.1
alpha-vantage>=3.0.0
black==25.1.0
prometheus_client==0.22.1
pandas<2.0.0,>=1.5.3
lm-format-enforcer==0.10.11
nest-asyncio==1.6.0
lark==1.2.2
Pygments==2.19.2
tzdata==2025.2
click<8.2,>=8.1.7
six==1.17.0
datasets==4.0.0
distro==1.9.0
torchvision==0.22.0
propcache==0.3.2
httptools==0.6.4
websockets==10.4
soupsieve==2.7
email_validator==2.2.0
sentencepiece==0.2.0
alpha_vantage==3.0.0
MarkupSafe==3.0.2
outlines_core==0.1.26
h11<0.15,>=0.13
pydantic_core==2.33.2
curl_cffi==0.11.4
typing-inspection==0.4.1
diskcache==5.6.3
regex==2024.11.6
uvicorn==0.35.0
scipy==1.16.0
typing_extensions==4.14.1
partial-json-parser==0.2.1.1.post6
cycler==0.12.1
tokenizers==0.21.2
aiohappyeyeballs==2.6.1
jsonschema-specifications==2025.4.1
astor==0.8.1
outlines==0.1.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
numba==0.61.2
fsspec==2025.3.0
rpds-py==0.26.0
mpmath==1.3.0
mlx-lm==0.26.0
pycountry==24.6.1
fonttools==4.58.5
tqdm==4.67.1
fastapi==0.116.0
pyarrow==20.0.0
platformdirs==4.3.8
fastapi-cli==0.0.8
torch==2.7.0
sentry-sdk==2.32.0
rignore==0.5.1
compressed-tensors==0.10.2
pathspec==0.12.1
sympy==1.14.0
mdurl==0.1.2
fastapi-cloud-cli==0.1.2
yarl==1.20.1
kiwisolver==1.4.8
pyparsing==3.2.3
referencing==0.36.2
certifi==2025.7.9
pydantic==2.11.7
vllm==0.9.2
httpcore==1.*
multidict==5.2.0
gguf==0.17.1
airportsdata==20250706
mistral_common==1.7.0
annotated-types==0.7.0
mypy_extensions==1.1.0
xxhash==3.5.0
pytz==2025.2
watchfiles==1.1.0
pybase64==1.4.1
pillow==11.3.0
starlette==0.46.2
anyio==4.9.0
packaging==24.2
sniffio==1.3.1
uvloop==0.21.0
pyzmq==27.0.0
iniconfig==2.1.0
frozendict==2.4.6
cffi==1.17.1
numpy<2.6,>=1.25.2